backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
pytest>=7.0.0
httpx>=0.24.0
google-genai>=1.32.0
elevenlabs>=0.2.0
python-multipart>=0.0.6
python-dotenv>=1.0.0